add static method convert CommentList to List<CommentResponse>

+++ MangaST_1_11/src/main/java/com/example/mangast/rating/RatingRepository.java
@@ -11,7 +11,7 @@ public interface RatingRepository extends JpaRepository<Rating, Integer> {
     SELECT AVG(fb.rate) FROM Rating fb
     WHERE fb.manga.id = :currentManga
     """)
-    float getAverageByManga(Manga currentManga);
+    double getAverageByManga(Manga currentManga);
 
     @Query("""
     SELECT fb FROM Rating fb
